[DX] Extract Job from batch bundle
