+++ spec/Pim/Bundle/BaseConnectorBundle/Archiver/ArchivableFileWriterArchiverSpec.php
@@ -5,7 +5,7 @@ namespace spec\Pim\Bundle\BaseConnectorBundle\Archiver;
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
 use Akeneo\Component\Batch\Item\ItemWriterInterface;
-use Akeneo\Bundle\BatchBundle\Job\Job;
+use Akeneo\Component\Batch\Job\Job;
 use Akeneo\Component\Batch\Step\AbstractStep;
 use Akeneo\Component\Batch\Step\ItemStep;
 use League\Flysystem\Adapter\Local as LocalAdapter;
