+++ src/Akeneo/Bundle/BatchBundle/Connector/ConnectorRegistry.php
@@ -3,10 +3,10 @@
 namespace Akeneo\Bundle\BatchBundle\Connector;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
-use Akeneo\Bundle\BatchBundle\Job\Job;
+use Akeneo\Component\Batch\Job\Job;
 use Akeneo\Bundle\BatchBundle\Job\JobFactory;
-use Akeneo\Bundle\BatchBundle\Step\StepFactory;
 use Akeneo\Component\Batch\Job\JobInterface;
+use Akeneo\Bundle\BatchBundle\Step\StepFactory;
 
 /**
  * Aims to register all connectors
@@ -38,7 +38,7 @@ class ConnectorRegistry
      * @param JobInstance $jobInstance
      *
      * @throws \LogicException
-     * @return \Akeneo\Component\Batch\Job\JobInterface
+     * @return JobInterface
      */
     public function getJob(JobInstance $jobInstance)
     {
