+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Connector/ConnectorRegistryTest.php
@@ -168,7 +168,7 @@ class ConnectorRegistryTest extends \PHPUnit_Framework_TestCase
         $logger = new Logger('JobLogger');
         $logger->pushHandler(new TestHandler());
 
-        return $this->getMock('Akeneo\\Bundle\\BatchBundle\\Job\\Job', array(), array('TestJob', $logger));
+        return $this->getMock('Akeneo\\Component\\Batch\\Job\\Job', array(), array('TestJob', $logger));
     }
 
     private function getConstructorDisabledMock($classname)
