+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Step/AbstractStepTest.php
@@ -2,9 +2,9 @@
 
 namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Job;
 
-use Akeneo\Bundle\BatchBundle\Job\Job;
 use Akeneo\Component\Batch\Job\BatchStatus;
 use Akeneo\Component\Batch\Job\ExitStatus;
+use Akeneo\Component\Batch\Job\Job;
 use Akeneo\Component\Batch\Job\JobInterruptedException;
 
 /**
